グラフデータ加工処理をフックにまとめた
